first refactoring
